Refacturation

+++ applications/girouette.c
@@ -1,21 +1,11 @@
-//_________________________________________________________
-// BINOME :
-// ETAT : 
-//__________________________________________________________
-
-//____ registres de périphériques du stm_32
 #include "stm32f10x.h"
 #include "../pilotes/clock/clock.h"
 #include "../pilotes/timer/timer.h"
 #include "../pilotes/gpio/gpio.h"
 #include "girouette.h"
 
-
-
-void Init_Girouette (void) {
-
-
-	// la girouette est rélier (physiquement) à PA5 ,PA6 et PA7
+void Init_Girouette(void) {
+	// la girouette est réliée (physiquement) à PA5 ,PA6 et PA7
 	// PA6= TIM3_CH1 et PA7=TIM3_CH2 donc les voie A et B du codeur sont bien reliés au timer
 	//Reste a configuré les PA5,6,7 en INPUT FLOATING
 	
@@ -32,32 +22,26 @@ void Init_Girouette (void) {
 	
 	//configuration de l'IT sur le PA5
 	(EXTI->IMR) = 0x01<<5 ;
-	
-		//configuration du type d'evenement : IT sur front montant
-		(EXTI->RTSR)|=(0x01<<5);
-		(EXTI->FTSR) &= ~(0x01 <<5);
+
+	//configuration du type d'evenement : IT sur front montant
+	(EXTI->RTSR)|=(0x01<<5);
+	(EXTI->FTSR) &= ~(0x01 <<5);
 	
 	//configuration du l'IT sur le M3
 	NVIC->ISER[0] |= NVIC_ISER_SETENA_23;
 	
 	// Priorité it NVIC
 	NVIC->IP[23]=4;
-		
-	
 }
 
-int Return_Angle_Girouette(void){
-	
-	int angle_girouette;
-	angle_girouette=(int)TIM3->CNT;
-	angle_girouette=angle_girouette/2;
-	return (angle_girouette);
+float Return_Angle_Girouette(void) {
+	float angle_girouette;
+	angle_girouette = (float)TIM3->CNT;
+	angle_girouette = angle_girouette / 2;
+	return angle_girouette;
 }
 
-void EXTI9_5_IRQHandler (void){
-	
+void EXTI9_5_IRQHandler(void) {
 	TIM3->CNT=0x0;
 	EXTI->PR |= 0x1 <<5;
-	
-	
 }
